@@ -1,84 +1,170 @@
-import click
-import ensemble.tools.constants as ct
+""" 
+Configurator
+====================================
+A class to store parameters for runtime execution
+"""
 
+# ============================================================================
+# STANDARD  IMPORTS
+# ============================================================================
+
+import click
+from dataclasses import dataclass, field
+import platform
+from typing import List, Any, Dict
+
+# ============================================================================
+# INTERNAL IMPORTS
+# ============================================================================
+
+
+from ensemble.tools.constants import (
+    DCT_DEFAULT_PATHS,
+    DCT_SIMULATORS,
+    DCT_RUNTIME_PARAM,
+)
 from ensemble.handler.symuvia import SymuviaConnector, SymuviaScenario
 from ensemble.handler.vissim.connector import VissimConnector, VissimScenario
-
 from ensemble.control.governor import MultiBrandPlatoonRegistry
+from ensemble.tools.screen import log_success, log_verify, log_warning
+
+# ============================================================================
+# CLASS AND DEFINITIONS
+# ============================================================================
 
 
-class Configurator(object):
-    """ 
-        This class stores some simulation configurations. 
+@dataclass
+class Configurator:
+    """ Configurator class for containing specific simulator parameter
+
+        Args:
+
+        verbose (bool):
+            Indicates if verbosity is required within the exit
+
+        info (bool):
+            Prints project information
+
+        platform (str):
+            Platform to run: Windows, Darwin, Linux
+
+        simulation_platform (str):
+            Traffic simulation platform: vissim, symuvia
+
+        scenario_files (list):
+            List of absolute files containing traffic scenarios
+
+        simulation_parameters (dict):
+            List of simulatio parameters. Check ``constants`` module for more information
     """
 
-    def __init__(self, verbose: bool = False) -> None:
-        import platform
+    verbose: bool = False
+    info: bool = True
+    platform: str = platform.system()
+    simulation_platform: str = ""
+    library_path: str = ""
 
-        self.verbose = verbose
-        self.platform = platform.system()
-        self.simulation_platform = ""
+    def __init__(self, **kwargs) -> None:
+        """ Configurator class for containing specific simulator parameter
+
+            Args:
+
+            verbose (bool):
+                Indicates if verbosity is required within the exit
+
+            info (bool):
+                Prints project information
+
+            platform (str):
+                Platform to run: Windows, Darwin, Linux
+
+            simulation_platform (str):
+                Traffic simulation platform: vissim, symuvia
+
+            scenario_files (list):
+                List of absolute files containing traffic scenarios
+
+            simulation_parameters (dict):
+                List of simulatio parameters. Check ``constants`` module for more information
+        """
+
+        for key, value in kwargs.items():
+            setattr(self, key, value)
+
         self.scenario_files = []
-        self.library_path = []
-        self.simulation_parameters = ct.DCT_RUNTIME_PARAM
+        self.simulation_parameters = DCT_RUNTIME_PARAM
 
     def set_simulation_platform(self, simulation_platform: str = "") -> None:
         """ A simpler setter for the simulation platform based on OS
 
-        :param simulation_platform: "symuvia" or "vissim", defaults to ""
-        :type simulation_platform: str, optional
+            Args:
+                simulation_platform (str): "symuvia" or "vissim", defaults to ""
         """
         if simulation_platform:
             self.simulation_platform = simulation_platform
             return
 
-        click.echo(click.style("Solving platform ", fg="blue", bold=True))
+        log_verify("Solving platform ", bold=True)
 
-        self.simulation_platform = ct.DCT_SIMULATORS.get(self.platform, "")
+        self.simulation_platform = DCT_SIMULATORS.get(self.platform, "")
 
         if self.verbose:
-            click.echo(
-                click.style(
-                    f"Setting default simulation platform platform {(self.  simulation_platform, self.platform)}",
-                    fg="yellow",
-                )
+            log_warning(
+                "Setting default simulation platform platform",
+                "\t{(self.simulation_platform, self.platform)}",
             )
 
-        self.library_path = ct.DCT_DEFAULT_PATHS.get((self.simulation_platform, self.platform))
+        key = (self.simulation_platform, self.platform)
+        self.library_path = DCT_DEFAULT_PATHS[key]
 
-        click.echo(click.style(f"Simulator path set to default value:\n \t {self.library_path}", fg="green",))
+        log_success(
+            "Simulator path set to default value:", f"\t{self.library_path}"
+        )
         return
 
     def update_values(self, **kwargs) -> None:
-        """ Configurator updater, pass a with keyword arguments to update
+        """ Configurator updater, pass a with keyword arguments to update. 
+            Just pass the desired parameter as a kewyword argument.
         """
 
         if kwargs.get("library_path"):
 
             self.library_path = kwargs.get("library_path", self.library_path)
 
-            click.echo(click.style(f"Setting new library path to user input:\n \t{self.library_path}", fg="yellow",))
+            log_verify(
+                "Setting new library path to user input:",
+                f"\t{self.library_path}",
+            )
 
         if kwargs.get("scenario_files"):
-            self.scenario_files = kwargs.get("scenario_files", self.scenario_files)
+            self.scenario_files = kwargs.get(
+                "scenario_files", self.scenario_files
+            )
 
-            click.echo(
-                click.style(f"Setting new scenario file(s) path to user input:  {self.library_path}", fg="yellow",)
+            log_verify(
+                "Setting new scenario file(s) path to user input:",
+                f"\t{self.scenario_files}",
             )
 
     def load_socket(self):
         """ Determines simulation platform to connect """
         if self.simulation_platform == "symuvia":
-            self.connector = SymuviaConnector(self.library_path, stepLaunchMode="traj")
+            self.connector = SymuviaConnector(
+                library_path=self.library_path, step_launch_mode="traj"
+            )
         else:
             self.connector = VissimConnector(self.library_path)
 
     def load_scenario(self):
         self.scenario_files = tuple(self.scenario_files)
         if self.simulation_platform == "symuvia":
-            scenario = SymuviaScenario.create_input(*self.scenario_files)  # expected input (fileA,fileB)
+            scenario = SymuviaScenario.create_input(
+                *self.scenario_files
+            )  # expected input (fileA,fileB)
         else:
-            scenario = VissimScenario.create_input(*self.scenario_files)  # expected input (fileA,fileB)
+            scenario = VissimScenario.create_input(
+                *self.scenario_files
+            )  # expected input (fileA,fileB)
 
         # Call connector (automatic dispatch)
         self.connector.load_scenario(scenario)
@@ -100,3 +186,7 @@ class Configurator(object):
     @property
     def total_steps(self):
         return self.simulation_parameters.get("total_steps")
+
+
+if __name__ == "__main__":
+    Configurator()
