@@ -1,10 +1,7 @@
 """
-    This module is able to receive the stream of data comming from the SymuVia platform and define a parser for a specific vehicle data suitable to perform platooning activities. 
-"""
-
-
-""" 
-    This module handles the Simulation response converting it into proper formats for querying data afterwards. 
+SymuVia Stream
+================
+This module is able to receive the stream of data comming from the SymuVia platform and define a parser for a specific vehicle data suitable to perform platooning activities. 
 """
 
 # ============================================================================
@@ -13,13 +10,15 @@
 
 from xmltodict import parse
 from xml.parsers.expat import ExpatError
+from ctypes import create_string_buffer
 
 # ============================================================================
 # INTERNAL IMPORTS
 # ============================================================================
 
-from ensemble.tools.stream import DataQuery
-from ensemble.component.vehicles import Vehicle, VehicleList
+from ensemble.metaclass.stream import DataQuery
+from symupy.utils.parser import vlists, response
+import ensemble.tools.constants as ct
 
 # ============================================================================
 # CLASS AND DEFINITIONS
@@ -27,233 +26,122 @@ from ensemble.component.vehicles import Vehicle, VehicleList
 
 
 class SimulatorRequest(DataQuery):
-    def __init__(self, channels):
-        super().__init__(channels)
-        self._strResponse = ""
-        self._vehList = []
-        self._dctData = {}
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self._str_response = create_string_buffer(ct.BUFFER_STRING)
 
-    def __repr__(self):
-        return f"{self.__class__.__name__}()"
-
-    def __str__(self):
-        return (
-            "Sim Time: {}, VehInNetwork: {}".format(self.current_time, self.current_nbveh)
-            if self.parsed_data
-            else "No vehicles detected"
-        )
-
-    # ============================================================================
+    # =========================================================================
     # MEMORY HANDLING
-    # ============================================================================
+    # =========================================================================
 
     @property
-    def query(self) -> str:
-        """
-            Parses response from simulator to data
-        """
-        return self._strResponse.value
+    def query(self):
+        """String response from the simulator"""
+        return self._str_response
 
     @query.setter
-    def query(self, response: str = None) -> None:
-        """
-            Parses response from simulator to data
-        """
-
-        # 1. gets string
-        self._strResponse = response
-
-        # 2. parses data
-        try:
-            self._dctData = parse(self._strResponse.value)
-        except ExpatError:
-            self._dctData = {}
-
-        # 3. vehicle list
-        self.update_vehicle_list()
+    def query(self, response: str):
+        self._str_response = response
+        for c in self._channels:
+            self.dispatch(c)
 
     @property
-    def parsed_data(self) -> dict:
-        """ 
-            Returns current data dictionary
-        """
-        return self._dctData
-
-    @property
-    def vehicles(self):
-        return self._vehList.vehicles
-
-    @property
-    def current_time(self) -> str:
-        return self.parsed_data.get("INST").get("@val")
+    def current_time(self) -> float:
+        return float(self.data_query.get("INST").get("@val"))
 
     @property
     def current_nbveh(self) -> int:
-        return self.parsed_data.get("INST").get("@nbVeh")
+        return int(self.data_query.get("INST").get("@nbVeh"))
 
-    # ============================================================================
+    @property
+    def data_query(self):
+        """ Direct parsing from the string buffer
+            
+            Returns:
+                simdata (OrderedDict): Simulator data parsed from XML
+        """
+        try:
+            dataveh = parse(self._str_response)
+            # Transform ordered dictionary into new keys
+            return dataveh
+        except ExpatError:
+            return {}
+        except AttributeError:
+            return {}
+
+    # =========================================================================
     # METHODS
-    # ============================================================================
+    # =========================================================================
 
-    def update_vehicle_list(self):
-        """ 
-            Construct and or update vehicle data
+    def get_vehicle_data(self) -> vlists:
+        """ Extracts vehicles information from simulators response
+
+            Returns:
+                t_veh_data (list): list of dictionaries containing vehicle data with correct formatting
+
         """
-        if self._vehList:
-            self._vehList.update_list(self.get_vehicle_data())
-            return
-        self._vehList = VehicleList.from_request(self.get_vehicle_data())
-
-    def get_vehicle_data(self) -> list:
-        """Extracts vehicles information from simulators response
-
-        :param response: Simulator response
-        :type response: str
-        :return: list of vehicles in the network
-        :rtype: list of dictionaries
-        """
-        if self.parsed_data.get("INST", {}).get("TRAJS") is not None:
-            veh_data = self.parsed_data.get("INST").get("TRAJS")
+        if self.data_query.get("INST", {}).get("TRAJS") is not None:
+            veh_data = self.data_query.get("INST").get("TRAJS")
             if isinstance(veh_data["TRAJ"], list):
-                return veh_data["TRAJ"]
-            return [veh_data["TRAJ"]]
+                return [SimulatorRequest.transform(d) for d in veh_data["TRAJ"]]
+            return [SimulatorRequest.transform(veh_data["TRAJ"])]
         return []
 
-    def get_vehicle_id(self) -> tuple:
-        """Extracts vehicle ids information from simulators response
+    @staticmethod
+    def transform(veh_data: dict):
+        """ Transform vehicle data from string format to coherent format
 
-        :return: tuple containing vehicle ids at current state in all network
-        :rtype: list
+            Args: 
+                veh_data (dict): vehicle data as received from simulator
+
+            Returns:
+                t_veh_data (dict): vehicle data with correct formatting 
+
+
+            Example: 
+                As an example, for an input of the following style ::
+
+                >>> v = OrderedDict([('@abs', '25.00'), ('@acc', '0.00'), ('@dst', '25.00'), ('@id', '0'), ('@ord', '0.00'), ('@tron', 'Zone_001'), ('@type', 'VL'), ('@vit', '25.00'), ('@voie', '1'),('@z', '0')])
+                >>> tv = SimulatorRequest.transform(v)
+                >>> # Transforms into 
+                >>> tv == {
+                >>>     "abscissa": 25.0,
+                >>>     "acceleration": 0.0,
+                >>>     "distance": 25.0,
+                >>>     "elevation": 0.0,
+                >>>     "lane": 1,
+                >>>     "link": "Zone_001",
+                >>>     "ordinate": 0.0,
+                >>>     "speed": 25.0,
+                >>>     "vehid": 0,
+                >>>     "vehtype": "VL",
+                >>> },
+
         """
-        return tuple(veh.get("@id") for veh in self.get_vehicle_data())
-
-    def query_vehicle_link(self, vehid: str, *args) -> tuple:
-        """ Extracts current vehicle link information from simulators response
-
-        :param vehid: vehicle id multiple arguments accepted
-        :type vehid: str
-        :return: vehicle link in tuple form
-        :rtype: tuple
-        """
-        vehids = set((vehid, *args)) if args else vehid
-        vehid_pos = self.query_vehicle_data_dict("@tron", vehids)
-        return tuple(vehid_pos.get(veh) for veh in vehids)
-
-    def query_vehicle_position(self, vehid: str, *args) -> tuple:
-        """ Extracts current vehicle distance information from simulators response
-
-        :param vehid: vehicle id multiple arguments accepted
-        :type vehid: str
-        :return: vehicle distance in link in tuple form
-        :rtype: tuple
-        """
-        vehids = set((vehid, *args)) if args else vehid
-        vehid_pos = self.query_vehicle_data_dict("@dst", vehids)
-        return tuple(vehid_pos.get(veh) for veh in vehids)
-
-    def query_vehicle_data_dict(self, dataval: str, vehid: str, *args) -> dict:
-        """ Extracts and filters vehicle data from the simulators response
-
-        :param dataval: parameter to be extracted e.g. '@id', '@dst'
-        :type dataval: str
-        :param vehid: vehicle id, multiple arguments accepted
-        :type vehid: str
-        :return: dictionary where key is @id and value is dataval
-        :rtype: dict
-        """
-        vehids = set((vehid, *args)) if args else set(vehid)
-        data_vehs = [(veh.get("@id"), veh.get(dataval)) for veh in self.get_vehicle_data() if veh.get("@id") in vehids]
-        return dict(data_vehs)
-
-    def is_vehicle_in_network(self, vehid: str, *args) -> bool:
-        """True if veh id is in the network at current state, for multiple arguments
-           True if all veh ids are in the network
-
-        :param vehid: Integer of vehicle id, comma separated if testing for multiple
-        :type vehid: int
-        :return: True if vehicle is in the network otherwise false
-        :rtype: bool
-        """
-        all_vehs = self.get_vehicle_id()
-        if not args:
-            return vehid in all_vehs
-        vehids = set((vehid, *args))
-        return set(vehids).issubset(set(all_vehs))
-
-    def vehicle_in_link(self, link: str, lane: str = "1") -> tuple:
-        """Returns a tuple containing vehicle ids traveling on the same link+lane at current state
-
-        :param link: link name
-        :type link: str
-        :param lane: lane number, defaults to '1'
-        :type lane: str, optional
-        :return: tuple containing vehicle ids
-        :rtype: tuple
-        """
-        return tuple(
-            veh.get("@id") for veh in self.get_vehicle_data() if veh.get("@tron") == link and veh.get("@voie") == lane
+        for key, val in veh_data.items():
+            response[ct.FIELD_DATA[key]] = ct.FIELD_FORMAT[key](val)
+        lkey = "@etat_pilotage"
+        response[ct.FIELD_DATA[lkey]] = ct.FIELD_FORMAT[lkey](
+            veh_data.get(lkey)
         )
+        return dict(response)
 
-    def is_vehicle_in_link(self, veh: str, link: str) -> bool:
-        """ Returns true if a vehicle is in a link at current state
-        
-        :param veh: vehicle id
-        :type veh: str
-        :param link: link name
-        :type link: str
-        :return: True if veh is in link
-        :rtype: bool
+    def is_vehicle_driven(self, vehid: int) -> bool:
+        """ Returns true if the vehicle state is exposed to a driven state
+
+            Args:
+                vehid (str):
+                    vehicle id
+            
+            Returns: 
+                driven (bool): True if veh is driven
         """
-        veh_ids = self.vehicle_in_link(link)
-        return set(veh).issubset(set(veh_ids))
+        if self.is_vehicle_in_network(vehid):
 
-    def vehicle_downstream_of(self, vehid: str) -> tuple:
-        """Get ids of vehicles downstream to vehid
-
-        :param vehid: integer describing id of reference veh
-        :type vehid: int
-        :return: tuple with ids of vehicles ahead (downstream)
-        :rtype: tuple
-        """
-        link = self.query_vehicle_link(vehid)[0]
-        vehpos = self.query_vehicle_position(vehid)[0]
-
-        vehids = set(self.vehicle_in_link(link))
-        neigh = vehids.difference(set(vehid))
-
-        neighpos = self.query_vehicle_position(*neigh)
-
-        return tuple(nbh for nbh, npos in zip(neigh, neighpos) if float(npos) > float(vehpos))
-
-    def vehicle_upstream_of(self, vehid: str) -> tuple:
-        """Get ids of vehicles upstream to vehid
-
-        :param vehid: integer describing id of reference veh
-        :type vehid: int
-        :return: tuple with ids of vehicles behind (upstream)
-        :rtype: tuple
-        """
-        link = self.query_vehicle_link(vehid)[0]
-        vehpos = self.query_vehicle_position(vehid)[0]
-
-        vehids = set(self.vehicle_in_link(link))
-        neigh = vehids.difference(set(vehid))
-
-        neighpos = self.query_vehicle_position(*neigh)
-
-        return tuple(nbh for nbh, npos in zip(neigh, neighpos) if float(npos) < float(vehpos))
-
-    def dispatch(self, channel: str) -> None:
-        """ This is a dispatcher for the vehicle
-
-        :param channel: Channel to broadcast to (FGC/RGC/ALL)
-        :type vehid: str
-
-        """
-
-        for subscriber, callback in self.get_subscribers(channel).items():
-            vehicle_env = {}
-            callback(vehicle_env)
-
-    def __contains__(self, elem: Vehicle) -> bool:
-        return elem in self._vehList
+            forced = tuple(
+                veh.get("driven") == True
+                for veh in self.get_vehicle_data()
+                if veh.get("vehid") == vehid
+            )
+            return any(forced)
+        return False
